Implemented Flot/Flotilla for Exposure Controller and view. Still need to clean up controller to remove old Open Flash Chart actions.
Had to remove OFC since it couldn't deal with null values very well. On the other hand, FLOT has
no axis titles, so I implemented them as a 2x3 table. Not great but it works. Still need to add
the curson tracking feature and also adapt the calculator (although this works very well with OFC,
so it can wait).

Signed-off-by: jmedding <jonmedding@gmail.com>